Refactor to single type and adapt app interface to frontend style and vice versa
