+++ src/dialogs/more-info/add-to.ts
@@ -1,0 +1,66 @@
+import type { HomeAssistant, TranslationDict } from "../../types";
+
+export interface EntityAddToAction {
+  /** Type of action. External is handled by external apps instead of in the frontend */
+  type: "default" | "external";
+  /** Whether the action is enabled and can be selected. */
+  enabled: boolean;
+  /** Translated name of the action */
+  name: string;
+  /** Optional translated description of the action */
+  description?: string;
+  /** MDI icon name (e.g., "mdi:car") */
+  icon: string;
+  /** Opaque payload for external action handling */
+  payload?: string;
+}
+
+export type EntityAddToActions = EntityAddToAction[];
+
+interface ActionDefinition {
+  translation_key: keyof TranslationDict["ui"]["dialogs"]["more_info_control"]["add_to"]["actions"];
+  icon: string;
+}
+
+export const DEFAULT_ACTION_DEFS: ActionDefinition[] = [
+  {
+    translation_key: "automation",
+    icon: "mdi:robot-outline",
+  },
+  {
+    translation_key: "script",
+    icon: "mdi:script-text-outline",
+  },
+];
+
+export const getDefaultAddToActions = (
+  hass: HomeAssistant,
+  entityId: string
+): EntityAddToActions =>
+  DEFAULT_ACTION_DEFS.map(
+    (def: ActionDefinition): EntityAddToAction => ({
+      type: "default",
+      enabled: true,
+      name: hass.localize(
+        `ui.dialogs.more_info_control.add_to.actions.${def.translation_key}`,
+        {
+          entity:
+            hass.states[entityId] !== undefined
+              ? hass.formatEntityName(hass.states[entityId], undefined)
+              : entityId,
+        }
+      ),
+      icon: def.icon,
+    })
+  );
+
+export function defaultActionHandler(
+  key: (typeof DEFAULT_ACTION_DEFS)[number]["translation_key"]
+) {
+  switch (key) {
+    case "automation":
+      break;
+    case "script":
+      break;
+  }
+}
